pyproject.toml
[tool.poetry]
name = "food"
version = "0.1.0"
description = ""
authors = ["og-tsarik <tsarikolya@gmail.com >"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
django-ckeditor = "^6.7.0"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^2023.10"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
